requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-static==0.0.6
Django==3.1.6
django-crispy-forms==1.10.0
django-jazzmin==2.4.4
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
telebot==0.0.4
urllib3==1.26.2
vine==5.0.0
vk-api==11.9.1
wcwidth==0.2.5
